@@ -1,7 +1,7 @@
 <?php
 /**
  * @link https://github.com/Izumi-kun/yii2-longpoll
- * @copyright Copyright (c) 2017 Viktor Khokhryakov
+ * @copyright Copyright (c) 2025 Viktor Khokhryakov
  * @license http://opensource.org/licenses/BSD-3-Clause
  */
 
@@ -21,7 +21,7 @@ class BaseLongPollAction extends Action
     /**
      * @var array
      */
-    public $events;
+    public array $events;
     /**
      * @var callable
      */
@@ -29,13 +29,12 @@ class BaseLongPollAction extends Action
     /**
      * @var string
      */
-    public $serverClass = Server::class;
+    public string $serverClass = Server::class;
 
     /**
      * @return Response
-     * @throws \yii\base\InvalidConfigException
      */
-    protected function runInternal()
+    protected function runInternal(): Response
     {
         /** @var Server $server */
         $server = Yii::createObject([
