@@ -1,7 +1,7 @@
 <?php
 /**
  * @link https://github.com/Izumi-kun/yii2-longpoll
- * @copyright Copyright (c) 2017 Viktor Khokhryakov
+ * @copyright Copyright (c) 2025 Viktor Khokhryakov
  * @license http://opensource.org/licenses/BSD-3-Clause
  */
 
@@ -33,27 +33,27 @@ use yii\web\View;
  */
 class LongPoll extends Widget
 {
-    public $url;
+    public string $url;
     /**
-     * @var EventCollectionInterface
+     * @var EventCollectionInterface|array|string
      */
-    public $eventCollection;
+    public EventCollectionInterface|array|string $eventCollection;
     /**
      * @var string event collection class name.
      */
-    public $eventCollectionClass = EventCollection::class;
+    public string $eventCollectionClass = EventCollection::class;
     /**
      * @var array additional options to be passed to JS registerer.
      */
-    public $clientOptions;
+    public array $clientOptions = [];
     /**
      * @var array params will be passed to JS XHR
      */
-    public $requestParams = [];
+    public array $requestParams = [];
     /**
      * @var string
      */
-    public $callback;
+    public string $callback;
 
     /**
      * @inheritdoc
@@ -79,7 +79,7 @@ class LongPoll extends Widget
         $view = $this->getView();
         LongPollAsset::register($view);
         $view->registerJs("jQuery.longpoll.register('$id', $options);", View::POS_END);
-        $view->registerJs("jQuery.longpoll.get('$id').start();", View::POS_READY);
+        $view->registerJs("jQuery.longpoll.get('$id').start();");
     }
 
     /**
@@ -88,7 +88,7 @@ class LongPoll extends Widget
      * @return array
      * @throws InvalidConfigException
      */
-    public static function createPollParams(EventCollectionInterface $eventCollection, $params = [])
+    public static function createPollParams(EventCollectionInterface $eventCollection, array $params = []): array
     {
         $events = [];
         foreach ($eventCollection->getEvents() as $event) {
@@ -116,7 +116,7 @@ class LongPoll extends Widget
         $options = $this->clientOptions;
         $options['url'] = Url::to($this->url);
         $options['params'] = self::createPollParams($this->eventCollection, $this->requestParams);
-        if ($this->callback) {
+        if (isset($this->callback)) {
             $options['callback'] = new JsExpression($this->callback);
         }
         return $options;
@@ -126,7 +126,7 @@ class LongPoll extends Widget
      * @param array $events
      * @throws InvalidConfigException
      */
-    public function setEvents($events)
+    public function setEvents(array $events)
     {
         $this->eventCollection = Yii::createObject([
             'class' => $this->eventCollectionClass,
