@@ -1,7 +1,7 @@
 <?php
 /**
  * @link https://github.com/Izumi-kun/yii2-longpoll
- * @copyright Copyright (c) 2017 Viktor Khokhryakov
+ * @copyright Copyright (c) 2025 Viktor Khokhryakov
  * @license http://opensource.org/licenses/BSD-3-Clause
  */
 
@@ -13,7 +13,7 @@ use yii\console\Controller;
 
 class MessageController extends Controller
 {
-    public $delay = 1;
+    public int $delay = 1;
     public $defaultAction = 'change';
 
     /**
