@@ -1,7 +1,7 @@
 <?php
 /**
  * @link https://github.com/Izumi-kun/yii2-longpoll
- * @copyright Copyright (c) 2017 Viktor Khokhryakov
+ * @copyright Copyright (c) 2025 Viktor Khokhryakov
  * @license http://opensource.org/licenses/BSD-3-Clause
  */
 
@@ -14,18 +14,18 @@ namespace izumi\longpoll;
 interface EventCollectionInterface
 {
     /**
-     * @param EventInterface|string $event the event object or key
+     * @param EventInterface|string|array $event the event object or key
      * @return EventInterface
      */
-    public function addEvent($event);
+    public function addEvent(EventInterface|string|array $event): EventInterface;
 
     /**
-     * @param EventInterface[]|array|string $events
+     * @param array|EventInterface[] $events
      */
-    public function setEvents($events);
+    public function setEvents(array $events): void;
 
     /**
      * @return EventInterface[] array of events (key => event).
      */
-    public function getEvents();
+    public function getEvents(): array;
 }
