@@ -1,7 +1,7 @@
 <?php
 /**
  * @link https://github.com/Izumi-kun/yii2-longpoll
- * @copyright Copyright (c) 2017 Viktor Khokhryakov
+ * @copyright Copyright (c) 2025 Viktor Khokhryakov
  * @license http://opensource.org/licenses/BSD-3-Clause
  */
 
@@ -37,10 +37,8 @@ use yii\web\Response;
  */
 class LongPollAction extends BaseLongPollAction
 {
-
     /**
      * @return Response
-     * @throws \yii\base\InvalidConfigException
      */
     public function run()
     {
