@@ -1,7 +1,7 @@
 <?php
 /**
  * @link https://github.com/Izumi-kun/yii2-longpoll
- * @copyright Copyright (c) 2017 Viktor Khokhryakov
+ * @copyright Copyright (c) 2025 Viktor Khokhryakov
  * @license http://opensource.org/licenses/BSD-3-Clause
  */
 
@@ -9,4 +9,11 @@ namespace tests;
 
 abstract class TestCase extends \PHPUnit\Framework\TestCase
 {
+    public function assertArraySubset(array $expected, array $actual)
+    {
+        foreach ($expected as $key => $value) {
+            $this->assertArrayHasKey($key, $actual);
+            $this->assertEquals($value, $actual[$key]);
+        }
+    }
 }
