@@ -1,7 +1,7 @@
 <?php
 /**
  * @link https://github.com/Izumi-kun/yii2-longpoll
- * @copyright Copyright (c) 2017 Viktor Khokhryakov
+ * @copyright Copyright (c) 2025 Viktor Khokhryakov
  * @license http://opensource.org/licenses/BSD-3-Clause
  */
 
@@ -21,7 +21,7 @@ class PollController extends Controller
             'index' => [
                 'class' => 'izumi\longpoll\LongPollAction',
                 'events' => ['newMessage'],
-                'callback' => [$this, 'longPollCallback'],
+                'callback' => $this->longPollCallback(...),
             ],
         ];
     }
