@@ -1,7 +1,7 @@
 <?php
 /**
  * @link https://github.com/Izumi-kun/yii2-longpoll
- * @copyright Copyright (c) 2017 Viktor Khokhryakov
+ * @copyright Copyright (c) 2025 Viktor Khokhryakov
  * @license http://opensource.org/licenses/BSD-3-Clause
  */
 
@@ -11,7 +11,7 @@ defined('YII_ENV') or define('YII_ENV', 'dev');
 require(__DIR__ . '/../../../vendor/autoload.php');
 require(__DIR__ . '/../../../vendor/yiisoft/yii2/Yii.php');
 
-Yii::setAlias('@tests', dirname(dirname(__DIR__)));
+Yii::setAlias('@tests', dirname(__DIR__, 2));
 
 \yii\helpers\FileHelper::removeDirectory(Yii::getAlias('@tests/app/runtime/logs'));
 
